feat: 로그인 안했을때 로직 추가 - 로그인 없이 앱 둘러보기 가능하도록 수정 - 로그인 안한경우 행사 페이지 제외하고 로그인 안내 문구 보여주기 refactor: 공통 스타일 컴포넌트 분리
