+++ src/components/Detail/InfoSkeleton.tsx
@@ -1,3 +1,4 @@
+import {SkeletonAnimation} from '@/styles/common';
 import styled, {css} from 'styled-components';
 
 function InfoSkeleton() {
@@ -30,21 +31,6 @@ function InfoSkeleton() {
     </SkeletonContainer>
   );
 }
-
-const SkeletonAnimation = css`
-  background: linear-gradient(90deg, #eeeeee 25%, #dddddd 37%, #eeeeee 63%);
-  background-size: 400% 100%;
-  animation: loading 1.4s ease infinite;
-
-  @keyframes loading {
-    0% {
-      background-position: 100% 0;
-    }
-    100% {
-      background-position: -100% 0;
-    }
-  }
-`;
 
 const SkeletonContainer = styled.div`
   display: flex;
