+++ src/components/common/Layout.tsx
@@ -27,7 +27,6 @@ const Container = styled.main<{
   display: flex;
   flex-direction: column;
   align-items: center;
-  justify-content: center;
   min-height: ${({$headerHeight, $footerHeight}) =>
     `calc(100vh - ${$headerHeight} - ${$footerHeight})`};
   margin-top: ${({$headerHeight}) => $headerHeight};
